recipes/recipe-06/azure-functions/requirements.txt
azure-functions
azure-identity
azure-mgmt-containerinstance